refactor: split sub-audit tasks into files
